requirements.txt
pydantic
fastapi
aiohttp
requests
pyjwt
python-jose[cryptography]